+++ src/components/Steps/index.ts
@@ -1,0 +1,3 @@
+export * from './constants';
+export { default as Steps } from './Steps.vue';
+export * from './types';